+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/listener/LogoutLinkListener.java
@@ -4,8 +4,8 @@ import javax.inject.Inject;
 
 import org.springframework.core.env.Environment;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.vaadin.spring.VaadinUIScope;
-import org.vaadin.spring.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.samples.mvp.ui.view.BannerView;
 
 import com.vaadin.server.VaadinServlet;
