+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/listener/ParticipantSelectedListener.java
@@ -2,8 +2,8 @@ package org.vaadin.spring.samples.mvp.ui.component.listener;
 
 import javax.inject.Inject;
 
-import org.vaadin.spring.VaadinUIScope;
-import org.vaadin.spring.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.EventBus;
 import org.vaadin.spring.events.EventBusScope;
 import org.vaadin.spring.events.EventScope;
