+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/selector/MarketDayPicker.java
@@ -8,7 +8,7 @@ import javax.inject.Inject;
 import org.apache.commons.lang3.ArrayUtils;
 import org.joda.time.DateTime;
 import org.springframework.core.env.Environment;
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 import org.vaadin.spring.samples.mvp.util.SSTimeUtil;
