+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/DSRUpdateHourlyView.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.lang3.math.NumberUtils;
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 import org.vaadin.spring.samples.mvp.dto.DSRUpdateHourlyDTO;
 import org.vaadin.spring.samples.mvp.ui.component.grid.HierarchicalDataGrid;
