+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/HeaderView.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import javax.annotation.PostConstruct;
 
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 import org.vaadin.spring.samples.mvp.ui.component.util.ControlsContext;
