+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/MainView.java
@@ -2,7 +2,7 @@ package org.vaadin.spring.samples.mvp.ui.view;
 
 import javax.annotation.PostConstruct;
 
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 
 import com.vaadin.navigator.View;
