+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/NavigationPanelView.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 import org.vaadin.spring.samples.mvp.ui.component.listener.NavigationPanelListener;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavContainerFactory;
