+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/ExecutionOperation1.java
@@ -16,8 +16,9 @@
 package org.vaadin.spring.samples.sidebar;
 
 import com.vaadin.ui.Notification;
+
 import org.springframework.stereotype.Component;
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.sidebar.SideBarItem;
 import org.vaadin.spring.sidebar.ThemeIcon;
