+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/PlanningOperation.java
@@ -17,8 +17,9 @@ package org.vaadin.spring.samples.sidebar;
 
 import com.vaadin.server.FontAwesome;
 import com.vaadin.ui.Notification;
+
 import org.springframework.stereotype.Component;
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.sidebar.FontAwesomeIcon;
 import org.vaadin.spring.sidebar.SideBarItem;
