+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/PlanningView1.java
@@ -20,7 +20,8 @@ import com.vaadin.navigator.ViewChangeListener;
 import com.vaadin.server.FontAwesome;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.VerticalLayout;
-import org.vaadin.spring.VaadinUIScope;
+
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.VaadinView;
 import org.vaadin.spring.sidebar.FontAwesomeIcon;
 import org.vaadin.spring.sidebar.SideBarItem;
