+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/SideBarUI.java
@@ -25,8 +25,9 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
 import com.vaadin.ui.themes.Reindeer;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.vaadin.spring.VaadinUI;
+import org.vaadin.spring.annotation.VaadinUI;
 import org.vaadin.spring.navigator.SpringViewProvider;
 import org.vaadin.spring.sidebar.SideBar;
