+++ spring-vaadin-mvp/src/test/java/org/vaadin/spring/navigator/FooView.java
@@ -1,6 +1,6 @@
 package org.vaadin.spring.navigator;
 
-import org.vaadin.spring.VaadinUIScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 
 import com.vaadin.navigator.View;
 import com.vaadin.navigator.ViewChangeListener.ViewChangeEvent;
