+++ spring-vaadin-mvp/src/test/java/org/vaadin/spring/navigator/TestConfig.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.navigator;
 
 import org.springframework.context.annotation.ComponentScan;
-import org.vaadin.spring.EnableVaadin;
+import org.vaadin.spring.annotation.EnableVaadin;
 
 @EnableVaadin
 @ComponentScan
