+++ spring-vaadin-security/src/test/java/org/vaadin/spring/security/test/DefaultTestConfiguration.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.security.test;
 
 import org.springframework.context.annotation.Configuration;
-import org.vaadin.spring.EnableVaadin;
+import org.vaadin.spring.annotation.EnableVaadin;
 
 @Configuration
 @EnableVaadin
