+++ spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/autowired/AnnotatedFooView.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.mvp.autowired;
 
 import org.vaadin.spring.UIScope;
-import org.vaadin.spring.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinComponent;
 import org.vaadin.spring.mvp.FooView;
 
 @UIScope
