+++ spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/autowired/AutowiredPresenter.java
@@ -2,7 +2,7 @@ package org.vaadin.spring.mvp.autowired;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.vaadin.spring.UIScope;
-import org.vaadin.spring.VaadinComponent;
+import org.vaadin.spring.annotation.VaadinComponent;
 import org.vaadin.spring.events.EventBus;
 import org.vaadin.spring.events.EventBusListenerMethod;
 import org.vaadin.spring.mvp.FooView;
