+++ spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/autowired/ScanConfig.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.mvp.autowired;
 
 import org.springframework.context.annotation.ComponentScan;
-import org.vaadin.spring.EnableVaadin;
+import org.vaadin.spring.annotation.EnableVaadin;
 
 @ComponentScan
 @EnableVaadin
