+++ spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/explicit/ExplicitConfig.java
@@ -3,8 +3,8 @@ package org.vaadin.spring.mvp.explicit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.vaadin.spring.EnableVaadin;
 import org.vaadin.spring.UIScope;
+import org.vaadin.spring.annotation.EnableVaadin;
 import org.vaadin.spring.events.EventBus;
 import org.vaadin.spring.mvp.FooView;
