+++ spring-vaadin-test/src/test/java/org/vaadin/spring/test/ExampleManualIntegrationTest.java
@@ -25,9 +25,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
-import org.vaadin.spring.EnableVaadin;
-import org.vaadin.spring.VaadinUIScope;
-import org.vaadin.spring.VaadinSessionScope;
+import org.vaadin.spring.annotation.EnableVaadin;
+import org.vaadin.spring.annotation.VaadinSessionScope;
+import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.EventBus;
 import org.vaadin.spring.events.EventBusScope;
 import org.vaadin.spring.events.EventScope;
