+++ spring-vaadin/src/main/java/org/vaadin/spring/annotation/VaadinUI.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.spring;
+package org.vaadin.spring.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
