+++ spring-vaadin/src/main/java/org/vaadin/spring/annotation/VaadinUIScope.java
@@ -13,22 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.spring;
+package org.vaadin.spring.annotation;
 
-import org.springframework.context.annotation.Import;
-import org.vaadin.spring.config.VaadinConfiguration;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-import java.lang.annotation.*;
+import org.springframework.context.annotation.Scope;
 
 /**
- * Brings in the machinery to setup Spring + Vaadin applications.
+ * Stereotype annotation for Spring's {@code @Scope("vaadin-ui")}.
  *
- * @author Josh Long (josh@joshlong.com)
  * @author Petter Holmström (petter@vaadin.com)
+ * @author Josh Long (josh@joshlong.com)
  */
-@Target(ElementType.TYPE)
+@Scope(org.vaadin.spring.internal.VaadinUIScope.VAADIN_UI_SCOPE_NAME)
+@Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Import(VaadinConfiguration.class)
-public @interface EnableVaadin {
+public @interface VaadinUIScope {
 }
