+++ spring-vaadin/src/main/java/org/vaadin/spring/internal/VaadinSessionScope.java
@@ -34,7 +34,7 @@ import org.springframework.beans.factory.config.Scope;
  * as the scope "{@value #VAADIN_SESSION_SCOPE_NAME}".
  *
  * @author Petter Holmström (petter@vaadin.com)
- * @see org.vaadin.spring.VaadinSessionScope
+ * @see org.vaadin.spring.annotation.VaadinSessionScope
  */
 public class VaadinSessionScope implements Scope, BeanFactoryPostProcessor {
