+++ spring-vaadin/src/main/java/org/vaadin/spring/navigator/VaadinView.java
@@ -16,7 +16,8 @@
 package org.vaadin.spring.navigator;
 
 import com.vaadin.ui.UI;
-import org.vaadin.spring.VaadinComponent;
+
+import org.vaadin.spring.annotation.VaadinComponent;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
@@ -28,7 +29,7 @@ import java.lang.annotation.Target;
  * <p>
  * This annotation is also a stereotype annotation, so Spring will automatically detect the annotated classes.
  * <b>However, the scope must be explicitly specified as the default singleton scope will not work!</b> You can use
- * the {@code prototype} scope or the {@link org.vaadin.spring.VaadinUIScope ui} scope.
+ * the {@code prototype} scope or the {@link org.vaadin.spring.annotation.VaadinUIScope ui} scope.
  * <p>
  * This is an example of a view that is mapped to an empty view name and is available for all UI subclasses in the application:
  *     <pre>
